Simplify implementation of the recent photo list.
Rework the implementation of the recent photo list so that instead of
trying to track the timestamp and sorting the array it just keeps the
array in order. When inserting, if there is a matching picture
information dicationary it is removed and the new photo is placed at the
beginning of the list. Also add the size limiting for the recent photo
list.